+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/connect-with-lines-form-model.js
@@ -1,9 +1,9 @@
 var _ = require('underscore');
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
-var template = require('./routing.tpl');
+var template = require('./connect-with-lines.tpl');
 var ColumnOptions = require('../column-options');
 
-var ROUTING_TYPES = require('./routing-types');
+var ROUTING_TYPES = require('./connect-with-lines-types');
 
 var TYPE_TO_META_MAP = {};
 ROUTING_TYPES.map(function (d) {
