+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/connect-with-lines-types.js
@@ -1,5 +1,5 @@
 module.exports = [
-  require('./routing/routing-to-single-point'),
-  require('./routing/routing-to-layer-all-to-all'),
-  require('./routing/routing-sequential')
+  require('./connect-with-lines/line-to-single-point'),
+  require('./connect-with-lines/line-to-layer-all-to-all'),
+  require('./connect-with-lines/line-sequential')
 ];
